refactor: adding down and sync to user walker and apply patches
patched to make sync and down working properly.
