+++ lib/commands/sync.js
@@ -4,24 +4,33 @@ var assert = require('./helper/assert.js');
 var migrationHook = require('./helper/migration-hook.js');
 
 module.exports = function (internals, config, callback) {
-  migrationHook(internals)
-    .then(function () {
-      var Migrator = require('../migrator.js');
-      var index = require('../../connect');
+  migrationHook(internals).then(function () {
+    var Migrator = require('../walker.js');
+    var index = require('../../connect');
 
-      if (!internals.argv.count) {
-        internals.argv.count = Number.MAX_VALUE;
-      }
-      index.connect({
+    if (!internals.argv.count) {
+      internals.argv.count = Number.MAX_VALUE;
+    }
+    index.connect(
+      {
         config: config.getCurrent().settings,
-        internals: internals
-      }, Migrator, function (err, migrator) {
+        internals: internals,
+        prefix: 'migration'
+      },
+      Migrator,
+      function (err, migrator) {
         if (!assert(err)) return;
 
         if (internals.locTitle) {
-          migrator.migrationsDir = path.resolve(internals.argv['migrations-dir'],
-            internals.locTitle);
-        } else { migrator.migrationsDir = path.resolve(internals.argv['migrations-dir']); }
+          migrator.migrationsDir = path.resolve(
+            internals.argv['migrations-dir'],
+            internals.locTitle
+          );
+        } else {
+          migrator.migrationsDir = path.resolve(
+            internals.argv['migrations-dir']
+          );
+        }
 
         internals.migrationsDir = migrator.migrationsDir;
 
@@ -29,9 +38,12 @@ module.exports = function (internals, config, callback) {
           if (!assert(err)) return;
           log.verbose('migration table created');
 
-          migrator.sync(internals.argv, internals.onComplete.bind(this,
-            migrator, internals, callback));
+          migrator.sync(
+            internals.argv,
+            internals.onComplete.bind(this, migrator, internals, callback)
+          );
         });
-      });
-    });
+      }
+    );
+  });
 };
