+++ lib/executors/versioned/v1.js
@@ -37,7 +37,9 @@ const execUnit = {
         return maybePromised(file, _file.down, [context.driver]);
       })
       .then(() => {
-        return Promise.promisify(context.deleteRecord.bind(context))(file);
+        return Promise.promisify(context.deleteMigrationRecord.bind(context))(
+          file
+        );
       })
       .then(context.driver.endMigration.bind(context.driver));
   }
