+++ lib/walker.js
@@ -60,14 +60,14 @@ Walker.prototype = {
   },
 
   deleteMigrationRecord: function (migration, callback) {
-    function onComplete (err) {
+    const onComplete = err => {
       if (err) {
         log.error(this.title + migration.name, err);
       } else {
         log.info(this.title + 'Processed', migration.name);
       }
       callback(err);
-    }
+    };
     this._driver.deleteMigration(
       this.internals.matching + '/' + migration.name,
       function (err) {
@@ -100,7 +100,7 @@ Walker.prototype = {
       .nodeify(callback);
   },
 
-  up: function ({ partialName, count }, callback) {
+  up: function ({ destination: partialName, count }, callback) {
     return Promise.all([
       File.loadFromFileystem(this.directory, this.prefix, this.internals),
       File.loadFromDatabase(
@@ -136,10 +136,10 @@ Walker.prototype = {
       .nodeify(callback);
   },
 
-  down: function ({ partialName, count }, callback) {
+  down: function ({ destination: partialName, count }, callback) {
     return File.loadFromDatabase(
       this.directory,
-      this.title,
+      this.prefix,
       this._driver,
       this.internals
     )
